feat: improve script
