+++ substrate_transfer_script/README.md
@@ -1,0 +1,17 @@
+# Substrate transfer precompile example scripts
+
+This piece of code exectues a transfer of 10 tokens from an EVM based account to a random substrate account.
+
+## Installation
+
+```sh
+npm i
+```
+
+## Run
+
+Make sure the node is running first on port `8545`
+
+```sh
+npm start
+```
